ISSUE#40 Исправлено: UserDetails перенесено с  ITUserDTO в ITUser
